Refactor, Add down_time, Add blocks_missed
